pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "wtconf"
version = "0.1.0"
description = "A tiny config launcher"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "slartibartmads" }]
dependencies = ["rich", "pyyaml"]

[project.scripts]
wtconf = "wtconf:main"